+++ client/src/features/advancedSearch/Group.tsx
@@ -103,10 +103,9 @@ const Group: React.FC<IGroup> = ({
   const labelForAria = conditionals[selectedKey]
   const ariaLabelForDropdowns = scopeToAriaLabel[parentScope]
 
-  console.log(selectedKey, bgColor)
   return (
     <div
-      className={`groupContainer ${bgColor} p-3 ${bgColor === 'bg-light' ? 'border' : ''} rounded-2`}
+      className={`groupContainer ${bgColor} p-3 ${nestedLevel !== 0 ? 'border' : ''} rounded-2`}
       style={{ minHeight: '100px' }}
     >
       <FormGroup>
@@ -205,7 +204,7 @@ const Group: React.FC<IGroup> = ({
                           parentScope={parentScope}
                           parentStateId={stateId}
                           parentGroupName={selectedKey}
-                          nestedLevel={nestedLevel}
+                          nestedLevel={nestedLevel + 1}
                           childInd={ind}
                           siblings={state}
                           parentBgColor={bgColor}